parisc: avoid WARNING: at kernel/cpu/idle.c:96
On Wed, 2013-05-08 at 22:42 +0200, Helge Deller wrote:
> We need to enable local irqs to avoid this runtime warning.
>
> Signed-off-by: Helge Deller <deller@gmx.de>
>
> diff --git a/arch/parisc/kernel/process.c b/arch/parisc/kernel/process.c
> index 55f92b6..2840d43 100644
> --- a/arch/parisc/kernel/process.c
> +++ b/arch/parisc/kernel/process.c
> @@ -286,3 +286,8 @@ void *dereference_function_descriptor(void *ptr)
>  	return ptr;
>  }
>  #endif
> +
> +void arch_cpu_idle(void)
> +{
> +	local_irq_enable();

This is definitely wrong.  we'd need at least a cpu_relax() in there.
Without that the cpu will bounce around in the idle loop.  I'd advise
that we simply keep the default cpu_idle_poll().

However, this looks like a kernel bug to me in cpu_idle_loop()?  not
something we should be working around in the arch code.

Thomas, what's going on here?  It looks like you can never avoid the
WARN_ON_ONCE if you don't provide any arch specific idle functions.  If
you wish to allow this case then the correct patch (keeping the WARN_ON)
would appear to be this one.

James

----
Reviewed-by: Srivatsa S. Bhat <srivatsa.bhat@linux.vnet.ibm.com>

Signed-off-by: Helge Deller <deller@gmx.de>