s390,sclp: add parameter to specify number of buffer pages
Add a kernel parameter to be able to specify the number of pages
to be used as output buffer by the line-mode sclp driver and the
vt220 sclp driver. The current number of output pages is 6, if the
service element is unavailable the boot messages alone can fill up
the output buffer. If this happens the system blocks until the
service element is working again. For a large LPAR with many
devices it is sensible to have the ability to increase the output
buffer size.

Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>